first commit in app1 branch

+++ src/users/users.service.ts
@@ -1,10 +1,12 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { Injectable, InternalServerErrorException, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { UserEntity } from 'src/typeorm/users.entity';
 import { Repository } from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { userStatus } from './valueobjects/user-Status.enum';
+import { UserDto } from './dto/user.dto';
+import { error } from 'console';
 
 @Injectable()
 export class UsersService {
@@ -14,54 +16,63 @@ export class UsersService {
     ){}
 
 
-    async returnUsers():Promise <Partial <UserEntity[]> > {
+    async getAllUsers():Promise <UserDto[]> {
         return await this.usersRepo.find()  
         
     }   
 
-    async returnUserById(id): Promise < Partial <UserEntity> > { 
+    async getUserById(id): Promise <UserDto>  { 
         const user = await this.usersRepo.createQueryBuilder("users")
-        .where("users.id = :id" , {id})
+        .where( "users.id = :id" , {id} )
         .getOne()
  
     // if (!user) {
     //    throw new NotFoundException(` user with id = ${id} not found `)
     // }
 
-     return user
+     return user.toUserDto()
 
     }
 
-    async creerUtilisateur( createUserDto: CreateUserDto ):Promise < Partial <UserEntity> >{
+    async createUser( createUserDto: CreateUserDto ):Promise <UserDto> {
         
        const  existUser = await this.usersRepo.createQueryBuilder("users")
        .where("users.email = :email" , {email : createUserDto.email})         
        .getOne()
 
        if( existUser ){
-        throw new NotFoundException(` l'utilisateur avec l'email =${createUserDto.email}  est deja existe `)
+        throw new error (` l'utilisateur avec l'email =${createUserDto.email}  est deja existe `)
        }else {
 
         const User = await this.usersRepo.save({ ...createUserDto });
 
-        const {id,password , ...newuser} = User
+        //const {id,password , ...newuser} = User
         
-        return newuser ; 
-
+        //return newuser ; 
+        return User.toUserDto();
        }
 
     } 
 
     async removeUser(id : any){
 
-        const user=this.returnUserById(id); 
-        await this.usersRepo.delete(id);
-        return " Suppression avec success ";
+        const user=this.getUserById(id); 
+        if (! user){        
+        throw new error(" user not existe ")    
+        }else{
+            try{
+            await this.usersRepo.delete(id);
+            return " Suppression avec success ";
+            }catch(e){
+            throw new InternalServerErrorException(" Un erreur se produit au cour de la suppression ")    
+            }
+        }
+
     }
 
-    async userUpdate( updateUserDto : UpdateUserDto  ,id : number ){
+    async userUpdate( updateUserDto : UpdateUserDto  ,id : number ):Promise<UserDto>{
 
-        const user = this.returnUserById(id); 
+        const user = this.getUserById(id); 
         if (user){
             const updateUser = await this.usersRepo.preload({
                 id ,
@@ -69,9 +80,10 @@ export class UsersService {
             })
 
             await this.usersRepo.save(updateUser)
-            const {id:number,password , ...user}= updateUser;
+           // const {id:number,password , ...user}= updateUser;
 
-            return user;
+            //return user;
+            return updateUser.toUserDto()
 
         }
 
@@ -79,7 +91,7 @@ export class UsersService {
     }
 
     async activateduser (id: number ){
-       const  user = await this.returnUserById(id);
+       const  user = await this.getUserById(id);
        if (!user){
         throw new NotFoundException( `User avec id =${id} est non disponible `);
        }
@@ -92,7 +104,7 @@ export class UsersService {
 
 
     async deactivateUser( id:number ){
-        const user= await this.returnUserById(id);
+        const user= await this.getUserById(id);
         if(!user){
             throw new NotFoundException( `User avec id =${id} est non disponible `);
         }
